+++ src/Infrastructure/Bundle/InfrastructureBundle/Migrations/Version20200211095359.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Migrations;
+
+use Doctrine\DBAL\Migrations\AbstractMigration;
+use Doctrine\DBAL\Schema\Schema;
+
+/**
+ * Add number_max_of_meetings_per_sheet to entity Type
+ */
+class Version20200211095359 extends AbstractMigration
+{
+    public function up(Schema $schema)
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE type ADD number_max_of_meetings_per_sheet INT DEFAULT NULL');
+    }
+
+    public function down(Schema $schema)
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE type DROP number_max_of_meetings_per_sheet');
+    }
+}