+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - VIMEET-2011 - Prioritize meeting request before planning
+- VIMEET-2047 - Add number max of meetings per sheet in participation type parameter
 - VIMEET-2046 - Limit upload image to 500ko
 - VIMEET-2049 - Fix nomenclature template object value: cast key to string
 ### Fixed